agregando recomendaciones

+++ app/Http/Controllers/LikeController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Like;
 use App\Exchange;
+use App\Talent;
 use Illuminate\Http\Request;
 
 use Illuminate\Support\Facades\Auth;
@@ -28,11 +29,22 @@ class LikeController extends Controller
      */
     public function create($canje_id)
     {
+
         $like = new Like;
         $like->exchange_id = $canje_id;
         $like->user_id = auth()->user()->id;
         $like->save();
     }
+
+    public function createTalent($talento_id)
+    {
+        
+        $like = new Like;
+        $like->talent_id = $talento_id;
+        $like->user_id = auth()->user()->id;
+        $like->save();
+    }
+
 
     /**
      * Store a newly created resource in storage.
@@ -81,20 +93,25 @@ class LikeController extends Controller
 
     public function changeLike(Request $request)
     {
-        $canje = Exchange::find($request->canje_id);
+        $tipo = $request->tipo;
 
-        if(($request->cambiar) == 1){
-            $like = Like::where('exchange_id', $request->canje_id)->where('user_id', Auth()->user()->id)->first();
-            if($like)
-            {
-                $this->destroy($like);
+        if($tipo == 'canje'){
+            $canje = Exchange::find($request->canje_id);
+            if(($request->cambiar) == 1){
+                    $like = Like::where('exchange_id', $request->canje_id)->where('user_id', Auth()->user()->id)->first();
+                if($like) $this->destroy($like);
+                else $this->create($request->canje_id);
             }
-            else
-            {
-                $this->create($request->canje_id);
+            return view('content.likes', compact('canje', 'tipo'));
+        }else{
+            $talent = Talent::find($request->talent_id);
+            if(($request->cambiar) == 1){
+                $like = Like::where('talent_id', $request->talent_id)->where('user_id', Auth()->user()->id)->first();
+                if($like) $this->destroy($like);
+                else $this->createTalent($request->talent_id);
             }
+            return view('content.likes', compact('talent', 'tipo'));
         }
-        return view('content.likes', compact('canje'));
     }
 
     /**
